pyproject.toml
[tool.poetry]
name = "wemarry pkg"
version = "0.1.0"
description = "WeMarry Packages"
authors = ["hyun <changja00@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 120

[tool.flake8]
ignore = ["E501", "E402", "E261", "E266"]
exclude = [".git", "__pycache__", "*/settings/*" ,"*/models.py"]
max-line-length = 120
count = true

[tool.isort]
atomic = "true"
profile = "black"

[tool.mypy]
strict = "True"
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
disallow_incomplete_defs = "True"
disallow_subclassing_any = "False"
disallow_untyped_decorators = "False"
check_untyped_defs = "True"
no_implicit_optional = "True"
explicit_package_bases = "True"
implicit_reexport = "True"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
django-extensions = "^3.2.3"
whitenoise = "^6.5.0"
django-ebhealthcheck = "^2.0.2"
pre-commit = "^3.3.3"
black = "^23.7.0"
mypy = "^1.5.1"
djangorestframework-dataclasses = "^1.3.0"
django-choices = "^2.0.0"
dacite = "^1.8.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
polyfactory = "^2.8.0"
factory-boy = "^3.3.0"
boto3 = "^1.28.44"
pillow = "^10.0.0"
django-storages = "^1.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



